clean up the tests
