+++ lib/bureaucrat/fields/ip_address_field.rb
@@ -1,0 +1,15 @@
+require 'bureaucrat/fields/char_field'
+
+module Bureaucrat
+  module Fields
+    class IPAddressField < CharField
+      def default_error_messages
+        super.merge(invalid: 'Enter a valid IPv4 address.')
+      end
+
+      def default_validators
+        [Validators::ValidateIPV4Address]
+      end
+    end
+  end
+end
