+++ test/fields/ip_address_field_test.rb
@@ -1,0 +1,17 @@
+require_relative '../test_helper'
+
+module IPAddressFieldTests
+  class Test_translation_errors < BureaucratTestCase
+    def setup
+      @field = Fields::IPAddressField.new({})
+    end
+
+    def test_translates_invalid_default
+      assert_equal(I18n.t('bureaucrat.default_errors.ip_address.invalid'), @field.error_messages[:invalid])
+    end
+
+    def test_translates_invalid_default
+      assert_equal(I18n.t('bureaucrat.default_errors.ip_address.invalid'), @field.error_messages[:invalid])
+    end
+  end
+end
